Preliminary support for "Lookup" on mobile.
Firefox for Android doesn't support context menus, so instead draw a CSS
bubble below the selected text.

The UI work is basically complete, but the text and href are dummy values.
Next step is figuring out how to plumb this into the contextMenus logic.

There is a lot of subtlety here regarding the position of the bubble and
its arrow, because the table, bubble, and selection are all dynamic,
but with AI assitance you can basically make CSS do anything.